// packages/hooks/src/useRequest/types.ts
export type UseRequestService<TData, TParams extends unknown[]> = (
  ...args: TParams
) => Promise<TData>;

export interface UseRequestFetchState<TData, TParams extends unknown[]> {
  loading: boolean;
  params?: TParams;
  data?: TData;
  error?: Error | unknown;
}

export type UseRequestOptions<TData, TParams extends unknown[], TPlugin> = {
  [K in keyof UseRequestBasicOptions<TData, TParams>]: UseRequestBasicOptions<TData, TParams>[K];
} & {
  [K in keyof TPlugin]: TPlugin[K];
};

export interface UseRequestPlugin<TData, TParams extends unknown[] = unknown[], TPlugin = any> {
  (
    fetchInstance: UseRequestFetch<TData, TParams>,
    options: UseRequestOptions<TData, TParams, TPlugin>,
  ): UseRequestPluginReturn<TData, TParams>;
  onInit?: (
    options: UseRequestOptions<TData, TParams, TPlugin>,
  ) => Partial<UseRequestFetchState<TData, TParams>>;
}

export interface UseRequestPluginReturn<TData, TParams extends unknown[]> {
  name?: string;
  onBefore?: (params: TParams) =>
    | ({
        stopNow?: boolean;
        returnNow?: boolean;
      } & Partial<UseRequestFetchState<TData, TParams>>)
    | void;

  onRequest?: (
    service: UseRequestService<TData, TParams>,
    params: TParams,
  ) => {
    servicePromise?: Promise<TData>;
  };

  onSuccess?: (data: TData, params: TParams) => void;
  onError?: (e: Error, params: TParams) => void;
  onFinally?: (params: TParams, data?: TData, e?: Error) => void;
  onCancel?: () => void;
  onMutate?: (data: TData) => void;
}
// ...
export interface useRequestResult<TData, TParams extends unknown[]> {
  /**
   * Is the service being executed.
   */
  loading: Readonly<Ref<boolean>>;

  /**
   * Data returned by service.
   */
  data: Readonly<Ref<TData | undefined>>;

  /**
   * 	Exception thrown by service.
   */
  error: Readonly<Ref<Error | undefined>>;

  /**
   * params	An array of parameters for the service being executed. For example, you triggered `run(1, 2, 3)`, then params is equal to `[1, 2, 3]`.
   */
  params: Readonly<Ref<TParams | []>>;

  /**
   * Ignore the current promise response.
   */
  cancel: UseRequestFetch<TData, TParams>['cancel'];

  /**
   * Use the last params, call `run` again.
   */
  refresh: UseRequestFetch<TData, TParams>['refresh'];

  /**
   * Use the last params, call `runAsync` again.
   */
  refreshAsync: UseRequestFetch<TData, TParams>['refreshAsync'];

  /**
   * Manually trigger the execution of the service, and the parameters will be passed to the service.
   */
  run: UseRequestFetch<TData, TParams>['run'];

  /**
   * Automatic handling of exceptions, feedback through `onError`
   */
  runAsync: UseRequestFetch<TData, TParams>['runAsync'];

  /**
   * Mutate `data` directly
   */
  mutate: UseRequestFetch<TData, TParams>['mutate'];
}